% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{cheby_coeffs}
\alias{cheby_coeffs}
\title{Chebyshev coefficients for a spectral kernel}
\usage{
cheby_coeffs(kernel, K, lmax, M = NULL, jackson = FALSE, cache = TRUE)
}
\arguments{
\item{kernel}{function of lambda}

\item{K}{polynomial order (>=1)}

\item{lmax}{spectral radius}

\item{M}{oversampling points (default 2*K)}
}
\description{
Chebyshev coefficients for a spectral kernel
}
\keyword{internal}
