% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{chebyshev_filter}
\alias{chebyshev_filter}
\title{Chebyshev polynomial filtering}
\usage{
chebyshev_filter(g, signal, coeffs, lambda_max_opt = NULL)
}
\arguments{
\item{g}{graph}

\item{signal}{numeric vector/matrix (n x m)}

\item{coeffs}{Chebyshev coefficients}

\item{lambda_max_opt}{optional precomputed lambda_max}
}
\description{
Chebyshev polynomial filtering
}
