% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{compute_differential_operator}
\alias{compute_differential_operator}
\title{Precompute differential operators}
\usage{
compute_differential_operator(g, oriented = TRUE)
}
\arguments{
\item{g}{gsp_graph}

\item{oriented}{logical; orientation used for incidence}
}
\value{
list with components \code{B} (incidence), \code{grad}, \code{div}
}
\description{
Returns a list containing the incidence matrix and closures for grad/div.
}
