% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{compute_log_scales}
\alias{compute_log_scales}
\title{Logarithmically spaced scales (PyGSP-compatible helper)}
\usage{
compute_log_scales(lmin, lmax, n_scales, t1 = 1, t2 = 2)
}
\arguments{
\item{lmin}{smallest non-zero eigenvalue proxy}

\item{lmax}{spectral radius}

\item{n_scales}{number of scales}

\item{t1}{lower kernel support parameter}

\item{t2}{upper kernel support parameter}
}
\description{
Logarithmically spaced scales (PyGSP-compatible helper)
}
\keyword{internal}
