% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{gabor_filter_bank}
\alias{gabor_filter_bank}
\title{Gabor filter bank}
\usage{
gabor_filter_bank(mus, sigma)
}
\arguments{
\item{mus}{center frequencies}

\item{sigma}{bandwidth}
}
\description{
Gabor filter bank
}
