% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{gft}
\alias{gft}
\title{Graph Fourier Transform}
\usage{
gft(g, signal, k = NULL)
}
\arguments{
\item{g}{graph}

\item{signal}{numeric vector (length n) or matrix (n x m)}

\item{k}{eigenpairs to use; NULL for full basis}
}
\description{
Graph Fourier Transform
}
