% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_airfoil}
\alias{graph_airfoil}
\title{Airfoil graph (PyGSP dataset)}
\usage{
graph_airfoil(normalized = FALSE)
}
\arguments{
\item{normalized}{logical; compute normalized Laplacian}
}
\value{
a \code{gsp_graph} with \code{coords}
}
\description{
Airfoil graph (PyGSP dataset)
}
