% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_barabasi_albert}
\alias{graph_barabasi_albert}
\title{Barabasi-Albert preferential attachment graph}
\usage{
graph_barabasi_albert(n, m = 2, seed = NULL, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{m}{edges per new node}

\item{seed}{optional RNG seed}

\item{normalized}{logical; normalized Laplacian}
}
\description{
Barabasi-Albert preferential attachment graph
}
