% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_bunny}
\alias{graph_bunny}
\title{Stanford Bunny graph (PyGSP dataset)}
\usage{
graph_bunny(normalized = FALSE)
}
\arguments{
\item{normalized}{logical; compute normalized Laplacian}
}
\value{
a \code{gsp_graph} with 3D \code{coords}
}
\description{
Stanford Bunny graph (PyGSP dataset)
}
