% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_cache_clear}
\alias{graph_cache_clear}
\title{Clear cached quantities stored in a graph.}
\usage{
graph_cache_clear(g)
}
\arguments{
\item{g}{a gsp_graph object}
}
\value{
the graph with cleared cache
}
\description{
Clear cached quantities stored in a graph.
}
