% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{graph_cache_invalidate_spectra}
\alias{graph_cache_invalidate_spectra}
\title{Invalidate cached spectra (eigenpairs, lambda_max, Chebyshev coeffs)}
\usage{
graph_cache_invalidate_spectra(g)
}
\arguments{
\item{g}{a gsp_graph object}
}
\value{
the graph with spectral cache cleared
}
\description{
Invalidate cached spectra (eigenpairs, lambda_max, Chebyshev coeffs)
}
