% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{graph_cache_invalidate_struct}
\alias{graph_cache_invalidate_struct}
\title{Invalidate cached Laplacians and degrees}
\usage{
graph_cache_invalidate_struct(g)
}
\arguments{
\item{g}{a gsp_graph object}
}
\value{
the graph with structural cache cleared
}
\description{
Invalidate cached Laplacians and degrees
}
