% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_complete}
\alias{graph_complete}
\title{Complete graph}
\usage{
graph_complete(n, directed = FALSE, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{directed}{logical; directed edges (no self-loops)}

\item{normalized}{logical; normalized Laplacian}
}
\description{
Complete graph
}
