% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_degree}
\alias{graph_degree}
\title{Get (and cache) degree vector.}
\usage{
graph_degree(g)
}
\arguments{
\item{g}{a gsp_graph object}
}
\value{
numeric vector of node degrees
}
\description{
Get (and cache) degree vector.
}
