% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{graph_eigenpairs}
\alias{graph_eigenpairs}
\title{Compute (and cache) Laplacian eigenpairs}
\usage{
graph_eigenpairs(g, k = NULL, which = "SM", normalized = g$normalized)
}
\arguments{
\item{g}{gsp_graph}

\item{k}{number of eigenpairs; NULL for full}

\item{which}{passed to RSpectra::eigs_sym when k < n (default "SM")}

\item{normalized}{use normalized Laplacian}
}
\description{
Compute (and cache) Laplacian eigenpairs
}
