% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_erdos_renyi}
\alias{graph_erdos_renyi}
\title{Erdos-Renyi G(n,p) random graph}
\usage{
graph_erdos_renyi(n, p, directed = FALSE, seed = NULL, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{p}{edge probability}

\item{directed}{logical; if TRUE, edges oriented independently}

\item{seed}{optional RNG seed}

\item{normalized}{logical; use normalized Laplacian}
}
\description{
Erdos-Renyi G(n,p) random graph
}
