% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_gradient}
\alias{graph_gradient}
\title{Weighted gradient operator on a graph}
\usage{
graph_gradient(g)
}
\arguments{
\item{g}{gsp_graph}
}
\value{
sparse matrix (n_edges x n_nodes)
}
\description{
Gradient matrix D such that Dx gives weighted edge differences (sqrt(w)*(x_i - x_j)).
}
