% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_grid2d}
\alias{graph_grid2d}
\title{2D grid graph (optionally torus)}
\usage{
graph_grid2d(nrow, ncol, periodic = FALSE, normalized = FALSE)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{periodic}{wrap edges to make a torus}

\item{normalized}{logical; compute normalized Laplacian}
}
\description{
2D grid graph (optionally torus)
}
