% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_grid3d}
\alias{graph_grid3d}
\title{3D grid graph (optionally torus)}
\usage{
graph_grid3d(nx, ny, nz, periodic = FALSE, normalized = FALSE)
}
\arguments{
\item{nx}{number of nodes in x}

\item{ny}{number of nodes in y}

\item{nz}{number of nodes in z}

\item{periodic}{wrap edges to make a 3D torus}

\item{normalized}{logical; compute normalized Laplacian}
}
\description{
3D grid graph (optionally torus)
}
