% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_laplacian}
\alias{graph_laplacian}
\title{Get (and cache) Laplacian.}
\usage{
graph_laplacian(g, normalized = g$normalized)
}
\arguments{
\item{g}{a gsp_graph object}

\item{normalized}{use normalized Laplacian}
}
\value{
sparse Laplacian matrix
}
\description{
Get (and cache) Laplacian.
}
