% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_logo}
\alias{graph_logo}
\title{GSP Logo graph (PyGSP dataset)}
\usage{
graph_logo(normalized = FALSE)
}
\arguments{
\item{normalized}{logical; compute normalized Laplacian}
}
\value{
a \code{gsp_graph} with \code{coords} and \code{info} indices
}
\description{
GSP Logo graph (PyGSP dataset)
}
