% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_minnesota}
\alias{graph_minnesota}
\title{Minnesota road network graph (PyGSP dataset)}
\usage{
graph_minnesota(connected = TRUE, normalized = FALSE)
}
\arguments{
\item{connected}{logical; if TRUE, match PyGSP's connected/binarized variant}

\item{normalized}{logical; compute normalized Laplacian}
}
\value{
a \code{gsp_graph} with \code{coords} and \code{labels}
}
\description{
Minnesota road network graph (PyGSP dataset)
}
