% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_path}
\alias{graph_path}
\title{Path graph (alias)}
\usage{
graph_path(n, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{normalized}{logical; compute normalized Laplacian}
}
\value{
a gsp_graph object
}
\description{
Path graph (alias)
}
