% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_random_geometric}
\alias{graph_random_geometric}
\title{Random geometric graph in 2D (unit square)}
\usage{
graph_random_geometric(n, radius, normalized = FALSE, seed = NULL)
}
\arguments{
\item{n}{nodes}

\item{radius}{connect nodes within this distance}

\item{normalized}{compute normalized Laplacian}

\item{seed}{optional RNG seed for reproducibility}
}
\value{
a gsp_graph object
}
\description{
Random geometric graph in 2D (unit square)
}
