% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_random_regular}
\alias{graph_random_regular}
\title{Random regular graph (configuration model)}
\usage{
graph_random_regular(n, degree, seed = NULL, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{degree}{degree for each node (n*degree must be even)}

\item{seed}{optional RNG seed}

\item{normalized}{logical; normalized Laplacian}
}
\description{
Random regular graph (configuration model)
}
