% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_rescale}
\alias{graph_rescale}
\title{Rescale edge weights by a factor (returns a new graph)}
\usage{
graph_rescale(g, factor)
}
\arguments{
\item{g}{a gsp_graph object}

\item{factor}{multiplicative factor for edge weights}
}
\value{
a new gsp_graph with rescaled weights
}
\description{
Rescale edge weights by a factor (returns a new graph)
}
