% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_ring}
\alias{graph_ring}
\title{Ring (or path) graph}
\usage{
graph_ring(n, periodic = TRUE, normalized = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{periodic}{if FALSE, returns a path graph}

\item{normalized}{logical; compute normalized Laplacian}
}
\description{
Ring (or path) graph
}
