% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_sbm}
\alias{graph_sbm}
\title{Stochastic Block Model graph}
\usage{
graph_sbm(block_sizes, P, normalized = FALSE, seed = NULL)
}
\arguments{
\item{block_sizes}{integer vector of community sizes}

\item{P}{matrix of edge probabilities (symmetric)}

\item{normalized}{compute normalized Laplacian}

\item{seed}{optional RNG seed for reproducibility}
}
\value{
a gsp_graph object
}
\description{
Stochastic Block Model graph
}
