% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_sparsify}
\alias{graph_sparsify}
\title{Sparsify small weights (in-place threshold)}
\usage{
graph_sparsify(g, tol = 1e-12)
}
\arguments{
\item{g}{a gsp_graph object}

\item{tol}{values with abs < tol are dropped}
}
\value{
a new gsp_graph with small weights removed
}
\description{
Sparsify small weights (in-place threshold)
}
