% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{heat_propagate}
\alias{heat_propagate}
\title{Heat diffusion of a signal for one or multiple times}
\usage{
heat_propagate(g, signal, t, K = 30, lmax = NULL)
}
\arguments{
\item{g}{graph}

\item{signal}{vector/matrix}

\item{t}{scalar or vector of times}

\item{K}{Chebyshev order}

\item{lmax}{optional spectral radius}
}
\value{
filtered signal(s), or list if multiple times
}
\description{
Heat diffusion of a signal for one or multiple times
}
