% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{igft}
\alias{igft}
\title{Inverse Graph Fourier Transform}
\usage{
igft(g, coeffs, k = NULL)
}
\arguments{
\item{g}{graph}

\item{coeffs}{spectral coefficients (k x m)}

\item{k}{eigenpairs to use; NULL for full basis}
}
\description{
Inverse Graph Fourier Transform
}
