% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{interpolate_laplacian}
\alias{interpolate_laplacian}
\title{Interpolation / inpainting of missing nodes via Laplacian smoothing}
\usage{
interpolate_laplacian(
  g,
  y,
  alpha = 0.01,
  tol = 1e-06,
  maxit = 1000,
  threads = NULL
)
}
\arguments{
\item{g}{graph}

\item{y}{vector of length n with NA for missing entries}

\item{alpha}{smoothing parameter}

\item{tol}{CG tolerance}

\item{maxit}{CG max iterations}

\item{threads}{optional number of threads (reserved)}
}
\description{
Interpolation / inpainting of missing nodes via Laplacian smoothing
}
