% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_exponential}
\alias{kernel_exponential}
\title{Exponential kernel exp(- (lambda / alpha)^beta)}
\usage{
kernel_exponential(alpha = 1, beta = 1)
}
\arguments{
\item{alpha}{scale parameter}

\item{beta}{shape parameter}
}
\value{
kernel function
}
\description{
Exponential kernel exp(- (lambda / alpha)^beta)
}
