% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_gabor}
\alias{kernel_gabor}
\title{Gabor kernel exp(-(lambda-mu)^2 / (2 sigma^2))}
\usage{
kernel_gabor(mu, sigma)
}
\arguments{
\item{mu}{center frequency}

\item{sigma}{bandwidth}
}
\value{
kernel function
}
\description{
Gabor kernel exp(-(lambda-mu)^2 / (2 sigma^2))
}
