% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_half_cosine}
\alias{kernel_half_cosine}
\title{Half-cosine transition band kernel
Smoothly transitions from 1 to 0 between low and high.}
\usage{
kernel_half_cosine(low, high)
}
\arguments{
\item{low}{lower cutoff frequency}

\item{high}{upper cutoff frequency}
}
\value{
kernel function
}
\description{
Half-cosine transition band kernel
Smoothly transitions from 1 to 0 between low and high.
}
