% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_mexican_hat}
\alias{kernel_mexican_hat}
\title{Mexican hat-like kernel lambda * exp(-lambda / sigma)}
\usage{
kernel_mexican_hat(sigma = 1)
}
\arguments{
\item{sigma}{scale parameter}
}
\value{
kernel function
}
\description{
Mexican hat-like kernel lambda * exp(-lambda / sigma)
}
