% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_meyer}
\alias{kernel_meyer}
\title{Smooth Meyer-like kernel (transition band)}
\usage{
kernel_meyer(low, high)
}
\arguments{
\item{low}{low cutoff}

\item{high}{high cutoff}
}
\description{
Smooth Meyer-like kernel (transition band)
}
