% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_rectangle}
\alias{kernel_rectangle}
\title{Rectangular pass-band}
\usage{
kernel_rectangle(low, high)
}
\arguments{
\item{low}{lower cutoff frequency}

\item{high}{upper cutoff frequency}
}
\value{
kernel function
}
\description{
Rectangular pass-band
}
