% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{kernel_wave}
\alias{kernel_wave}
\title{Wave (cosine) kernel for wave equation propagation
h(lambda) = cos(t * sqrt(lambda))}
\usage{
kernel_wave(t)
}
\arguments{
\item{t}{time parameter}
}
\description{
Wave (cosine) kernel for wave equation propagation
h(lambda) = cos(t * sqrt(lambda))
}
