% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{kron_reduction}
\alias{kron_reduction}
\title{Kron reduction (Schur complement) of a graph Laplacian}
\usage{
kron_reduction(g, keep)
}
\arguments{
\item{g}{gsp_graph}

\item{keep}{integer vector of node indices to keep (1-based)}
}
\value{
reduced gsp_graph on |keep| nodes
}
\description{
Reduces a graph by eliminating a subset of nodes via Schur complement.
}
