% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{lambda_max}
\alias{lambda_max}
\title{Estimate largest eigenvalue (spectral radius)}
\usage{
lambda_max(g, normalized = g$normalized)
}
\arguments{
\item{g}{a gsp_graph object}

\item{normalized}{logical; use normalized Laplacian}
}
\value{
the largest eigenvalue
}
\description{
Estimate largest eigenvalue (spectral radius)
}
