% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{layout_circle}
\alias{layout_circle}
\title{Circle/Ring Layout}
\usage{
layout_circle(n)
}
\arguments{
\item{n}{number of nodes}
}
\value{
matrix of 2D coordinates
}
\description{
Circle/Ring Layout
}
\keyword{internal}
