% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{layout_grid}
\alias{layout_grid}
\title{Grid Layout (for grid graphs)}
\usage{
layout_grid(g)
}
\arguments{
\item{g}{graph object}
}
\value{
matrix of coordinates
}
\description{
Grid Layout (for grid graphs)
}
\keyword{internal}
