% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{layout_random}
\alias{layout_random}
\title{Random Layout}
\usage{
layout_random(n, dim = 2)
}
\arguments{
\item{n}{number of nodes}

\item{dim}{2 or 3}
}
\value{
matrix of coordinates
}
\description{
Random Layout
}
\keyword{internal}
