% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{layout_spectral}
\alias{layout_spectral}
\title{Spectral Layout (Laplacian Eigenmap)}
\usage{
layout_spectral(g, dim = 2)
}
\arguments{
\item{g}{graph object}

\item{dim}{2 or 3}
}
\value{
matrix of coordinates
}
\description{
Spectral Layout (Laplacian Eigenmap)
}
\keyword{internal}
