% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{layout_spring}
\alias{layout_spring}
\title{Fruchterman-Reingold Spring Layout}
\usage{
layout_spring(g, dim = 2, iterations = 50, k = NULL, temp = 0.1)
}
\arguments{
\item{g}{graph object}

\item{dim}{2 or 3}

\item{iterations}{number of iterations (default: 50)}

\item{k}{optimal distance between nodes (default: auto)}

\item{temp}{initial temperature (default: 0.1)}
}
\value{
matrix of coordinates
}
\description{
Force-directed graph layout using the Fruchterman-Reingold algorithm.
}
\keyword{internal}
