% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{legend_position}
\alias{legend_position}
\title{Legend Position Helper}
\usage{
legend_position(position = "right")
}
\arguments{
\item{position}{one of "right", "left", "top", "bottom", "none"}
}
\value{
ggplot2 theme element for legend position
}
\description{
Convenience function to adjust legend position.
}
\examples{
\dontrun{
g <- graph_ring(20)
g <- graph_set_coords(g, "circle")
plot_graph(g, signal = rnorm(20)) + legend_position("bottom")
}

}
