% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mexican_hat_wavelet_bank}
\alias{mexican_hat_wavelet_bank}
\title{Mexican-hat wavelet bank}
\usage{
mexican_hat_wavelet_bank(scales, n_scales = NULL, lmax = NULL, lpfactor = 20)
}
\arguments{
\item{scales}{scales vector, or a gsp_graph}

\item{n_scales}{number of scales (if scales is a graph)}

\item{lmax}{spectral radius}

\item{lpfactor}{factor for low-frequency bound}
}
\value{
list of kernel functions
}
\description{
Mexican-hat wavelet bank
}
