% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{meyer_wavelet_bank}
\alias{meyer_wavelet_bank}
\title{Meyer wavelet bank (lowpass + bandpasses)}
\usage{
meyer_wavelet_bank(scales = NULL, lmax = NULL, n_scales = NULL)
}
\arguments{
\item{scales}{multiplicative scales (e.g., 2^(0:J-1)), or a gsp_graph}

\item{lmax}{spectral radius}

\item{n_scales}{number of scales (if scales is NULL or a graph)}
}
\value{
list of kernel functions
}
\description{
Meyer wavelet bank (lowpass + bandpasses)
}
