% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{modulation_transform}
\alias{modulation_transform}
\title{Modulation / Gabor vertex-frequency transform}
\usage{
modulation_transform(
  g,
  signal,
  mus,
  sigma,
  K = 30,
  lmax = NULL,
  jackson = FALSE
)
}
\arguments{
\item{g}{graph}

\item{signal}{vector or matrix (n x m)}

\item{mus}{vector of center frequencies}

\item{sigma}{bandwidth of gabor kernel}

\item{K}{Chebyshev order}

\item{lmax}{optional spectral radius}

\item{jackson}{logical; use Jackson damping}
}
\value{
list of filtered signals (one per mu)
}
\description{
Modulation / Gabor vertex-frequency transform
}
