% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{new_graph}
\alias{new_graph}
\title{Create a graph object from adjacency}
\usage{
new_graph(adjacency, coords = NULL, normalized = FALSE, directed = FALSE)
}
\arguments{
\item{adjacency}{sparse adjacency (dgCMatrix)}

\item{coords}{optional node coordinates (matrix/data.frame)}

\item{normalized}{logical; if TRUE compute normalized Laplacian}

\item{directed}{logical; if TRUE, Laplacian uses directed form (random-walk)}
}
\description{
Create a graph object from adjacency
}
