% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_band_energy}
\alias{plot_band_energy}
\title{Plot Band Energy Distribution}
\usage{
plot_band_energy(W, title = "Wavelet Energy by Scale")
}
\arguments{
\item{W}{sgwt_coeffs object from sgwt()}

\item{title}{plot title}
}
\value{
ggplot object
}
\description{
Bar plot showing energy distribution across wavelet scales.
}
\examples{
\dontrun{
g <- graph_sensor(50, seed = 42)
W <- sgwt(g, rnorm(50))
plot_band_energy(W)
}

}
