% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot a Graph Signal}
\usage{
plot_signal(g, signal, ...)
}
\arguments{
\item{g}{gsp_graph object}

\item{signal}{numeric vector of signal values}

\item{...}{additional arguments passed to plot_graph}
}
\value{
ggplot object
}
\description{
Quick visualization of a signal on a graph. Wrapper around plot_graph.
}
\examples{
\dontrun{
g <- graph_sensor(100, seed = 42)
signal <- rnorm(100)
plot_signal(g, signal)
}

}
