% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_transform.R
\name{print.dsgwt_riesz}
\alias{print.dsgwt_riesz}
\title{Print method for dsgwt_riesz}
\usage{
\method{print}{dsgwt_riesz}(x, ...)
}
\arguments{
\item{x}{dsgwt_riesz object}

\item{...}{additional arguments (ignored)}
}
\value{
invisibly returns the object
}
\description{
Print method for dsgwt_riesz
}
