% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_transform.R
\name{print.dsgwt_steer}
\alias{print.dsgwt_steer}
\title{Print method for dsgwt_steer}
\usage{
\method{print}{dsgwt_steer}(x, ...)
}
\arguments{
\item{x}{dsgwt_steer object}

\item{...}{additional arguments (ignored)}
}
\value{
invisibly returns the object
}
\description{
Print method for dsgwt_steer
}
