% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{print.gsp_graph}
\alias{print.gsp_graph}
\title{Print method for gsp_graph}
\usage{
\method{print}{gsp_graph}(x, ...)
}
\arguments{
\item{x}{a gsp_graph object}

\item{...}{additional arguments (ignored)}
}
\value{
invisibly returns the graph
}
\description{
Print method for gsp_graph
}
