% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_transform.R
\name{print.sgwt_coeffs}
\alias{print.sgwt_coeffs}
\title{Print SGWT Coefficients}
\usage{
\method{print}{sgwt_coeffs}(x, ...)
}
\arguments{
\item{x}{sgwt_coeffs object}

\item{...}{ignored}
}
\description{
Print SGWT Coefficients
}
