% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{prox_l1}
\alias{prox_l1}
\title{L1 soft-threshold (proximal)}
\usage{
prox_l1(x, tau)
}
\arguments{
\item{x}{vector or matrix}

\item{tau}{threshold parameter}
}
\description{
L1 soft-threshold (proximal)
}
