% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{prox_tv1d}
\alias{prox_tv1d}
\title{1D total variation proximal}
\usage{
prox_tv1d(y, lambda)
}
\arguments{
\item{y}{numeric vector}

\item{lambda}{weight}
}
\description{
1D total variation proximal
}
