% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{pyramid_analysis}
\alias{pyramid_analysis}
\title{Pyramid analysis (coarse + detail) via simple decimation}
\usage{
pyramid_analysis(g, signal, keep = NULL)
}
\arguments{
\item{g}{gsp_graph}

\item{signal}{numeric vector length n}

\item{keep}{indices kept (if NULL, keep every other node)}
}
\value{
list with coarse, detail, keep, elim
}
\description{
Pyramid analysis (coarse + detail) via simple decimation
}
