% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{pyramid_synthesis}
\alias{pyramid_synthesis}
\title{Pyramid synthesis (reconstruction) from coarse/detail}
\usage{
pyramid_synthesis(p)
}
\arguments{
\item{p}{list from pyramid_analysis}
}
\value{
reconstructed signal
}
\description{
Pyramid synthesis (reconstruction) from coarse/detail
}
