% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{regression_tikhonov}
\alias{regression_tikhonov}
\title{Graph Tikhonov regression}
\usage{
regression_tikhonov(g, y, tau, tol = 1e-06, maxit = 1000, threads = NULL)
}
\arguments{
\item{g}{graph}

\item{y}{observed labels/signals (n or n x m)}

\item{tau}{regularization strength}

\item{tol}{solver tolerance}

\item{maxit}{max iterations}

\item{threads}{optional threads}
}
\description{
Solves min_x 0.5||y - x||^2 + tau x^T L x
}
