% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsp-package.R
\docType{package}
\name{rgsp-package}
\alias{rgsp}
\alias{rgsp-package}
\title{rgsp: Graph Signal Processing in R (Port of PyGSP)}
\description{
Core graph signal processing algorithms with C++ acceleration via RcppArmadillo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourname/rgsp}
}

}
\author{
\strong{Maintainer}: Your Name \email{you@example.com}

}
\keyword{internal}
