% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scale_fill_gsp}
\alias{scale_fill_gsp}
\title{Fill Scale Presets for Heatmaps}
\usage{
scale_fill_gsp(palette = "viridis", direction = 1)
}
\arguments{
\item{palette}{one of "viridis", "spectral", "magma", "inferno", "plasma"}

\item{direction}{1 for normal direction, -1 for reversed}
}
\value{
ggplot2 scale object
}
\description{
Get pre-defined fill scales for spectrogram and coefficient visualizations.
}
\examples{
\dontrun{
g <- graph_ring(12)
W <- sgwt(g, rnorm(12))
plot_wavelet_coeffs(W) + scale_fill_gsp("magma")
}

}
