% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_transform.R
\name{summary.sgwt_coeffs}
\alias{summary.sgwt_coeffs}
\title{Summarize Wavelet Coefficients}
\usage{
\method{summary}{sgwt_coeffs}(object, ...)
}
\arguments{
\item{object}{sgwt_coeffs object}

\item{...}{ignored}
}
\description{
Summarize Wavelet Coefficients
}
