% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{tight_frame_regular}
\alias{tight_frame_regular}
\alias{tight_frame_held}
\alias{tight_frame_simoncelli}
\alias{tight_frame_papadakis}
\title{Tight frame bank presets (Regular/Held/Simoncelli/Papadakis style)}
\usage{
tight_frame_regular(lmax, J = 3)

tight_frame_held(lmax, J = 3)

tight_frame_simoncelli(lmax, J = 3)

tight_frame_papadakis(lmax, J = 3)
}
\arguments{
\item{lmax}{spectral radius}

\item{J}{number of dyadic bands}
}
\description{
Tight frame bank presets (Regular/Held/Simoncelli/Papadakis style)
}
