% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{tikhonov_smooth}
\alias{tikhonov_smooth}
\title{Tikhonov smoothing: solve (I + tau L) x = b}
\usage{
tikhonov_smooth(g, b, tau, tol = 1e-06, maxit = 1000, threads = NULL)
}
\arguments{
\item{g}{graph}

\item{b}{vector or matrix of observations}

\item{tau}{smoothing strength}

\item{tol}{CG tolerance}

\item{maxit}{CG max iterations}

\item{threads}{optional number of threads (reserved for future parallel solver)}
}
\description{
Tikhonov smoothing: solve (I + tau L) x = b
}
