% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{tv_inpaint}
\alias{tv_inpaint}
\title{TV-based inpainting (missing entries) using primal-dual prox}
\usage{
tv_inpaint(g, y, lambda, maxit = 300, tau = 0.5, sigma = 0.5, tol = 1e-06)
}
\arguments{
\item{g}{graph}

\item{y}{length-n vector with NA for missing entries}

\item{lambda}{TV weight}

\item{maxit}{iterations}

\item{tau}{primal step}

\item{sigma}{dual step}

\item{tol}{stopping tolerance}
}
\description{
TV-based inpainting (missing entries) using primal-dual prox
}
