% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{validate_signal_dims}
\alias{validate_signal_dims}
\title{Validate signal dimensions against graph}
\usage{
validate_signal_dims(g, signal)
}
\arguments{
\item{g}{graph}

\item{signal}{vector or matrix}
}
\description{
Validate signal dimensions against graph
}
\keyword{internal}
