% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wavelet_heat_bank}
\alias{wavelet_heat_bank}
\title{Heat wavelet bank at given scales}
\usage{
wavelet_heat_bank(scales)
}
\arguments{
\item{scales}{numeric vector of scales}
}
\description{
Heat wavelet bank at given scales
}
