% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wavelet_heat_transform}
\alias{wavelet_heat_transform}
\title{Compute heat wavelet transform}
\usage{
wavelet_heat_transform(g, signal, scales, K = 30, lmax = NULL, jackson = FALSE)
}
\arguments{
\item{g}{graph}

\item{signal}{vector/matrix}

\item{scales}{scales for heat kernels}

\item{K}{Chebyshev order for filtering}

\item{lmax}{optional spectral radius}

\item{jackson}{logical; apply Jackson damping}
}
\value{
list of filtered signals, one per scale
}
\description{
Compute heat wavelet transform
}
